requirements.txt
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.5
Pillow==7.1.2
pykwalify==1.7.0
pyparsing==2.4.7
pyradiomics==3.0
python-dateutil==2.8.1
PyWavelets==1.0.0
PyYAML==5.3.1
radiomics==0.1
scikit-image==0.17.2
scipy==1.4.1
SimpleITK==1.2.4
six==1.15.0
tifffile==2020.6.3
